+++ bibtexautocomplete/bibtex/matching.py
@@ -7,8 +7,26 @@ from typing import Set, Tuple
 from .entry import BibtexEntry
 from .normalize import normalize_str, normalize_str_weak
 
-ENTRY_CERTAIN_MATCH = 1000  # Maximum score
-ENTRY_NO_MATCH = 0  # Minimum score
+CERTAIN_MATCH = 1000  # Any score above is certain
+NO_MATCH = 0  # Any score below is no match
+
+# Matching score, XOR-ed together to get final score
+# Higher score = more likely to match
+
+# Added if DOIs match
+MATCH_DOI = CERTAIN_MATCH
+
+# Added if title match (after normalization)
+MATCH_TITLE = CERTAIN_MATCH // 2
+MATCH_TITLE_WEAK = CERTAIN_MATCH // 4  # uses a stronger normalization algorithm
+
+# Full author list match,
+MATCH_AUTHORS_FULL = CERTAIN_MATCH // 2
+MATCH_AUTHORS_INCLUDED = CERTAIN_MATCH // 4  # A.authors included in B.authors or B in A
+MATCH_AUTHORS_INTERSECTS = CERTAIN_MATCH // 8  # Non-empty intersection
+
+# source.year == query.year
+MATCH_YEAR = CERTAIN_MATCH // 4
 
 
 def author_set(entry: BibtexEntry) -> Set[str]:
@@ -38,26 +56,31 @@ def match_score(a: BibtexEntry, b: BibtexEntry) -> int:
     """
     if a.doi is not None and a.doi == b.doi:
         # DOIs are unique => entries match
-        return ENTRY_CERTAIN_MATCH
+        return MATCH_DOI
     if a.title is not None and b.title is not None:
         # Title is our best identifier
         if normalize_str_weak(a.title) == normalize_str_weak(b.title):
-            score = ENTRY_CERTAIN_MATCH // 2
+            score = MATCH_TITLE
         elif normalize_str(a.title) == normalize_str(b.title):
-            score = ENTRY_CERTAIN_MATCH // 3
+            score = MATCH_TITLE_WEAK
         else:
-            return ENTRY_NO_MATCH
+            return NO_MATCH
         authors_common, authors_a, authors_b = common_authors(a, b)
         if authors_common == 0 and authors_a > 0 and authors_b > 0:
             # No common authors despite some authors being known on both sides
-            return ENTRY_NO_MATCH
+            return NO_MATCH
         if authors_common != 0:
             if authors_common == authors_a and authors_common == authors_b:
-                score += ENTRY_CERTAIN_MATCH // 3
+                score += MATCH_AUTHORS_FULL
             elif authors_common == authors_a or authors_common == authors_b:
-                score += ENTRY_CERTAIN_MATCH // 6
+                score += MATCH_AUTHORS_INCLUDED
             else:
-                score += ENTRY_CERTAIN_MATCH // 9
+                score += MATCH_AUTHORS_INTERSECTS
+        if a.year is not None and b.year is not None:
+            if a.year != b.year:
+                # Different years
+                return NO_MATCH
+            score += MATCH_YEAR
         # Could check additional fields here
         return score
-    return ENTRY_NO_MATCH
+    return NO_MATCH