+++ CHANGELOG.md
@@ -1,11 +1,15 @@
 # Change Log
 
+## Version ??? - Not yet released
+
+- Add year check to match algorithm, entry with different year will not match
+
+
 ## Version 1.2.1 - 2023-05-12
 
 - Add `-w --overwrite` and `-W --dont-overwrite` flags
 - Fix flag `-f` being ignored in query pre-condition checks, leading
   to skipping queries even if new data could be obtained.
-
 ## Version 1.2.0 - 2023-04-14
 
 - Add semantic scholar lookup